rcracers-313/requirements.txt
numpy 
casadi 
pyglet==1.5.28
matplotlib
opencv-python
alpaqa==1.0.0a1
pyserial
simple-pyspin
rich
pynput
bleak
